requirements.txt
openai==0.27.6
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.18.1
